+++ .github/workflows/release-notes.md
@@ -4,8 +4,9 @@
 - Requires K-indexer 0.1.14
 
 ### New features
-- Added following/followers count on "Profile"
-- Added user's following/followers list accessible from "Profile" or user's page; removed "Followed" page
+- Added following/followers/blocked count on "Profile" page
+- Added access to "Following/Followers" page on "Profile" page and "User" page; removed "Followed" page
+- Moved access to "Blocked users" page from left sidebar to "Profile" page
 - Added back button on all pages
 - Added "Follow" button on users list
 - Cleaned all users lists by removing timestamp